% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan.R
\name{img_team_logo}
\alias{img_team_logo}
\alias{img_teamcolors_hex}
\title{Embed logos}
\usage{
img_team_logo(pattern, ...)

img_teamcolors_hex(...)
}
\arguments{
\item{pattern}{regular expression matching names passed to 
\code{\link[dplyr]{filter}}}

\item{...}{arguments passed to \code{\link[htmltools]{img}}}
}
\description{
Embed logos
}
\examples{
img_team_logo("New York Mets", width = 64)
img_teamcolors_hex(width = 64)
}
