% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{league_palette}
\alias{league_palette}
\title{ggplot2 palettes}
\usage{
league_palette(lg, which = 1)
}
\arguments{
\item{lg}{character vector for the league identifier}

\item{which}{Which set of colors do you want? Default is 1 for "primary"}
}
\description{
ggplot2 palettes
}
\seealso{
teamcolors
}
