% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan.R
\name{mono_light_team}
\alias{mono_light_team}
\alias{mono_dark_team}
\alias{duo_team}
\alias{duo_accent_team}
\alias{duo_accent_inverse_team}
\title{Xaringan themes}
\usage{
mono_light_team(pattern, colors = c(1, 2), ...)

mono_dark_team(pattern, colors = c(1, 2), ...)

duo_team(pattern, colors = c(1, 2), ...)

duo_accent_team(pattern, colors = c(1, 2), ...)

duo_accent_inverse_team(pattern, colors = c(1, 2), ...)
}
\arguments{
\item{pattern}{regular expression matched to \code{name} 
variable from \code{teamcolors}}

\item{colors}{vector of colors to choose. Default is \code{c(1, 2)}
for primary and secondary colors.}

\item{...}{arguments passed to \code{xaringanthemer} functions}
}
\description{
Xaringan themes
}
\seealso{
\code{\link[xaringanthemer]{mono_light}}

\code{\link[xaringanthemer]{mono_dark}}

\code{\link[xaringanthemer]{duo}}

\code{\link[xaringanthemer]{duo_accent}}

\code{\link[xaringanthemer]{duo_accent_inverse}}
}
