% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_sport_col}
\alias{show_sport_col}
\title{Displays palettes for all teams for a specified sport}
\usage{
show_sport_col(sport, ...)
}
\arguments{
\item{sport}{character vector (basketball, soccer, football, hockey)}

\item{...}{arguments passed to other functions}
}
\description{
Displays palettes for all teams for a specified sport
}
\examples{
show_sport_col(sport = "soccer")
}
\seealso{
\code{\link[scales]{show_col}}
}
