% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teamcolors}
\alias{teamcolors}
\title{Color palettes for professional sports teams}
\format{A data frame with one row for each professional team and five variables:
\describe{
 \item{name}{the name of the team}
 \item{primary}{the team's primary color}
 \item{secondary}{the team's secondary color}
 \item{tertiary}{the team's tertiary color}
 \item{quaternary}{the team's quaternary color}
 \item{sport}{the league in which the team plays}
}}
\source{
\url{https://github.com/teamcolors/teamcolors.github.io/tree/master/src/scripts/data/leagues}
}
\usage{
teamcolors
}
\description{
Color palettes for professional sports teams
}
\details{
The colors given are HTML hexidecimal values. See \code{\link[grDevices]{colors}}
for more information.
}
\examples{
data(teamcolors)

if (require(Lahman) & require(dplyr)) {
  pythag <- Teams \%>\%
    filter(yearID == 2014) \%>\%
    select(name, W, L, R, RA) \%>\%
    mutate(wpct = W / (W+L), exp_wpct = 1 / (1 + (RA/R)^2)) \%>\%
    # St. Louis Cardinals do not match
    left_join(teamcolors, by = "name")
  with(pythag, plot(wpct, exp_wpct, bg = primary, col = secondary, pch = 21, cex = 3))
}
if (require(ggplot2)) {
  primary_cols <- pythag$primary
  names(primary_cols) <- pythag$name
  ggplot(data = pythag, aes(x = exp_wpct, y = wpct)) + 
    geom_point(shape = 21, size = 5) + 
    scale_color_manual(values = secondary) + 
    scale_fill_manual(values = primary_cols)
}

}
\keyword{datasets}

